move auth_jwt modules to /lib. (force travis re-build) https://github.com/ideaq/time/issues/70
