+++ lib/auth_basic_validate.js
